assignment_evaluator/requirements.txt
Flask
SQLAlchemy
groq
python-dotenv
